refactor(islamic-website): use proper filename, move all get data functions to utils
