+++ apps/islamic-website/app/news/page.tsx
@@ -1,5 +1,8 @@
 import { NewsPage } from '@haqq/islamic-website/news-page';
 import { getNewsPageContent } from '../../utils/get-news-data';
+import { REVALIDATE_TIME } from '../../constants';
+
+export const revalidate = REVALIDATE_TIME;
 
 export const metadata = {
   title: 'IslamicCoin | News',
