+++ apps/islamic-website/app/partnerships/page.tsx
@@ -1,4 +1,4 @@
-import { DEPLOY_URL } from '../../contansts';
+import { DEPLOY_URL } from '../../constants';
 import { Metadata } from 'next';
 
 export { PartnershipPage as default } from '@haqq/islamic-website/partnership-page';
