+++ apps/islamic-website/app/whitepaper/page.tsx
@@ -1,27 +1,9 @@
-import { storyblokInit, apiPlugin } from '@storyblok/js';
 import { WhitepaperPage } from '@haqq/islamic-website/whitepaper-page';
-import { DEPLOY_URL } from '../../contansts';
+import { DEPLOY_URL, REVALIDATE_TIME } from '../../constants';
 import { Metadata } from 'next';
+import { getWhitepaperContent } from '../../utils/get-whitepaper-content';
 
-const STORYBLOK_ACCESS_TOKEN = process.env['STORYBLOK_ACCESS_TOKEN'];
-const VERCEL_ENV = process.env['VERCEL_ENV'];
-
-async function getWhitepaperContent() {
-  const { storyblokApi } = storyblokInit({
-    accessToken: STORYBLOK_ACCESS_TOKEN,
-    use: [apiPlugin],
-  });
-
-  if (!storyblokApi) {
-    throw new Error('Failed to init storyblok');
-  }
-
-  const response = await storyblokApi.get('cdn/stories/wp', {
-    version: VERCEL_ENV === 'production' ? 'published' : 'draft',
-  });
-
-  return response.data.story.content.body;
-}
+export const revalidate = REVALIDATE_TIME;
 
 export const metadata: Metadata = {
   title: 'IslamicCoin | Whitepaper',
