+++ apps/islamic-website/constants.ts
